+++ src/wireviz/DataClasses.py
@@ -13,16 +13,17 @@ class Connector:
     manufacturer: Optional[str] = None
     manufacturer_part_number: Optional[str] = None
     internal_part_number: Optional[str] = None
+    style: Optional[str] = None
     category: Optional[str] = None
     type: Optional[str] = None
     subtype: Optional[str] = None
     pincount: Optional[int] = None
     notes: Optional[str] = None
-    pinout: List[Any] = field(default_factory=list)
-    pinnumbers: List[Any] = field(default_factory=list)
+    pinlabels: List[Any] = field(default_factory=list)
+    pins: List[Any] = field(default_factory=list)
     color: Optional[str] = None
-    show_name: bool = True
-    show_pincount: bool = True
+    show_name: bool = None
+    show_pincount: bool = None
     hide_disconnected_pins: bool = False
     autogenerate: bool = False
     loops: List[Any] = field(default_factory=list)
@@ -32,28 +33,37 @@ class Connector:
         self.ports_right = False
         self.visible_pins = {}
 
+        if self.style == 'simple':
+            if self.pincount and self.pincount > 1:
+                raise Exception('Connectors with style set to simple may only have one pin')
+            self.pincount = 1
+
         if self.pincount is None:
-            if self.pinout:
-                self.pincount = len(self.pinout)
-            elif self.pinnumbers:
-                self.pincount = len(self.pinnumbers)
-            elif self.category == 'ferrule':
-                self.pincount = 1
+            if self.pinlabels:
+                self.pincount = len(self.pinlabels)
+            elif self.pins:
+                self.pincount = len(self.pins)
             else:
-                raise Exception('You need to specify at least one, pincount, pinout or pinnumbers')
+                raise Exception('You need to specify at least one, pincount, pins or pinlabels')
 
-        if self.pinout and self.pinnumbers:
-            if len(self.pinout) != len(self.pinnumbers):
-                raise Exception('Given pinout and pinnumbers size mismatch')
+        if self.pinlabels and self.pins:
+            if len(self.pinlabels) != len(self.pins):
+                raise Exception('Given pins and pinlabels size mismatch')
 
-        # create default lists for pinnumbers (sequential) and pinouts (blank) if not specified
-        if not self.pinnumbers:
-            self.pinnumbers = list(range(1, self.pincount + 1))
-        if not self.pinout:
-            self.pinout = [''] * self.pincount
+        # create default lists for pins (sequential) and pinlabels (blank) if not specified
+        if not self.pins:
+            self.pins = list(range(1, self.pincount + 1))
+        if not self.pinlabels:
+            self.pinlabels = [''] * self.pincount
 
-        if len(self.pinnumbers) != len(set(self.pinnumbers)):
-            raise Exception('Pin numbers are not unique')
+        if len(self.pins) != len(set(self.pins)):
+            raise Exception('Pins are not unique')
+
+        if self.show_name is None:
+            self.show_name = not self.autogenerate # hide auto-generated designators by default
+
+        if self.show_pincount is None:
+            self.show_pincount = self.style != 'simple' # hide pincount for simple (1 pin) connectors by default
 
         for loop in self.loops:
             # TODO: check that pins to connect actually exist