+++ src/wireviz/Harness.py
@@ -3,13 +3,15 @@
 
 from wireviz.DataClasses import Connector, Cable
 from graphviz import Graph
-from wireviz import wv_colors
+from wireviz import wv_colors, wv_helper
+from wireviz.wv_colors import get_color_hex
 from wireviz.wv_helper import awg_equiv, mm2_equiv, tuplelist2tsv, \
     nested_html_table, flatten2d, index_if_list, html_line_breaks, \
-    graphviz_line_breaks, remove_line_breaks
+    graphviz_line_breaks, remove_line_breaks, open_file_read, open_file_write
 from collections import Counter
 from typing import List
 from pathlib import Path
+import re
 
 
 class Harness:
@@ -33,23 +35,23 @@ class Harness:
         for (name, pin) in zip([from_name, to_name], [from_pin, to_pin]):  # check from and to connectors
             if name is not None and name in self.connectors:
                 connector = self.connectors[name]
-                if pin in connector.pinnumbers and pin in connector.pinout:
-                    if connector.pinnumbers.index(pin) == connector.pinout.index(pin):
+                if pin in connector.pins and pin in connector.pinlabels:
+                    if connector.pins.index(pin) == connector.pinlabels.index(pin):
                         # TODO: Maybe issue a warning? It's not worthy of an exception if it's unambiguous, but maybe risky?
                         pass
                     else:
-                        raise Exception(f'{name}:{pin} is defined both in pinout and pinnumbers, for different pins.')
-                if pin in connector.pinout:
-                    if connector.pinout.count(pin) > 1:
+                        raise Exception(f'{name}:{pin} is defined both in pinlabels and pins, for different pins.')
+                if pin in connector.pinlabels:
+                    if connector.pinlabels.count(pin) > 1:
                         raise Exception(f'{name}:{pin} is defined more than once.')
                     else:
-                        index = connector.pinout.index(pin)
-                        pin = connector.pinnumbers[index] # map pin name to pin number
+                        index = connector.pinlabels.index(pin)
+                        pin = connector.pins[index] # map pin name to pin number
                         if name == from_name:
                             from_pin = pin
                         if name == to_name:
                             to_pin = pin
-                if not pin in connector.pinnumbers:
+                if not pin in connector.pins:
                     raise Exception(f'{name}:{pin} not found.')
 
         self.cables[via_name].connect(from_name, from_pin, via_pin, to_name, to_pin)
@@ -77,51 +79,41 @@ class Harness:
 
         # prepare ports on connectors depending on which side they will connect
         for _, cable in self.cables.items():
-            for connection in cable.connections:
-                if connection.from_port is not None:  # connect to left
-                    self.connectors[connection.from_name].ports_right = True
-                if connection.to_port is not None:  # connect to right
-                    self.connectors[connection.to_name].ports_left = True
+            for connection_color in cable.connections:
+                if connection_color.from_port is not None:  # connect to left
+                    self.connectors[connection_color.from_name].ports_right = True
+                if connection_color.to_port is not None:  # connect to right
+                    self.connectors[connection_color.to_name].ports_left = True
 
         for key, connector in self.connectors.items():
-            if connector.category == 'ferrule':
 
-                rows = [[connector.manufacturer,
-                        f'MPN: {connector.manufacturer_part_number}' if connector.manufacturer_part_number else None,
-                        f'IPN: {connector.internal_part_number}' if connector.internal_part_number else None],
-                        [html_line_breaks(connector.type), html_line_breaks(connector.subtype), connector.color, '<!-- colorbar -->' if connector.color else None],
-                        [html_line_breaks(connector.notes)]]
-                html = nested_html_table(rows)
+            rows = [[connector.name if connector.show_name else None],
+                    [connector.manufacturer,
+                     f'MPN: {connector.manufacturer_part_number}' if connector.manufacturer_part_number else None,
+                     f'IPN: {connector.internal_part_number}' if connector.internal_part_number else None],
+                    [html_line_breaks(connector.type),
+                     html_line_breaks(connector.subtype),
+                     f'{connector.pincount}-pin' if connector.show_pincount else None,
+                     connector.color, '<!-- colorbar -->' if connector.color else None],
+                    '<!-- connector table -->' if connector.style != 'simple' else None,
+                    [html_line_breaks(connector.notes)]]
+            html = nested_html_table(rows)
 
-                if connector.color: # add color bar next to color info, if present
-                    colorbar = f' bgcolor="{wv_colors.translate_color(connector.color, "HEX")}" width="4"></td>' # leave out '<td' from string to preserve any existing attributes of the <td> tag
-                    html = html.replace('><!-- colorbar --></td>', colorbar)
+            if connector.color: # add color bar next to color info, if present
+                colorbar = f' bgcolor="{wv_colors.translate_color(connector.color, "HEX")}" width="4"></td>' # leave out '<td' from string to preserve any existing attributes of the <td> tag
+                html = html.replace('><!-- colorbar --></td>', colorbar)
 
-                dot.node(key, label=f'<{html}>', shape='none', margin='0', style='filled', fillcolor='white')
-
-            else:  # not a ferrule
-
-                rows = [[connector.name if connector.show_name else None],
-                        [connector.manufacturer,
-                         f'MPN: {connector.manufacturer_part_number}' if connector.manufacturer_part_number else None,
-                         f'IPN: {connector.internal_part_number}' if connector.internal_part_number else None],
-                        [html_line_breaks(connector.type),
-                         html_line_breaks(connector.subtype),
-                         f'{connector.pincount}-pin' if connector.show_pincount else None],
-                        '<!-- connector table -->',
-                        [html_line_breaks(connector.notes)]]
-                html = nested_html_table(rows)
-
-                pinouts = []
-                for pinnumber, pinname in zip(connector.pinnumbers, connector.pinout):
-                    if connector.hide_disconnected_pins and not connector.visible_pins.get(pinnumber, False):
+            if connector.style != 'simple':
+                pinlist = []
+                for pin, pinlabel in zip(connector.pins, connector.pinlabels):
+                    if connector.hide_disconnected_pins and not connector.visible_pins.get(pin, False):
                         continue
-                    pinouts.append([f'<td port="p{pinnumber}l">{pinnumber}</td>' if connector.ports_left else None,
-                                    f'<td>{pinname}</td>' if pinname else '',
-                                    f'<td port="p{pinnumber}r">{pinnumber}</td>' if connector.ports_right else None])
+                    pinlist.append([f'<td port="p{pin}l">{pin}</td>' if connector.ports_left else None,
+                                    f'<td>{pinlabel}</td>' if pinlabel else '',
+                                    f'<td port="p{pin}r">{pin}</td>' if connector.ports_right else None])
 
                 pinhtml = '<table border="0" cellspacing="0" cellpadding="3" cellborder="1">'
-                for i, pin in enumerate(pinouts):
+                for i, pin in enumerate(pinlist):
                     pinhtml = f'{pinhtml}<tr>'
                     for column in pin:
                         if column is not None:
@@ -130,21 +122,22 @@ class Harness:
                 pinhtml = f'{pinhtml}</table>'
                 html = html.replace('<!-- connector table -->', pinhtml)
 
-                dot.node(key, label=f'<{html}>', shape='none', margin='0', style='filled', fillcolor='white')
 
-                if len(connector.loops) > 0:
-                    dot.attr('edge', color='#000000:#ffffff:#000000')
-                    if connector.ports_left:
-                        loop_side = 'l'
-                        loop_dir = 'w'
-                    elif connector.ports_right:
-                        loop_side = 'r'
-                        loop_dir = 'e'
-                    else:
-                        raise Exception('No side for loops')
-                    for loop in connector.loops:
-                        dot.edge(f'{connector.name}:p{loop[0]}{loop_side}:{loop_dir}',
-                                 f'{connector.name}:p{loop[1]}{loop_side}:{loop_dir}')
+            dot.node(key, label=f'<{html}>', shape='none', margin='0', style='filled', fillcolor='white')
+
+            if len(connector.loops) > 0:
+                dot.attr('edge', color='#000000:#ffffff:#000000')
+                if connector.ports_left:
+                    loop_side = 'l'
+                    loop_dir = 'w'
+                elif connector.ports_right:
+                    loop_side = 'r'
+                    loop_dir = 'e'
+                else:
+                    raise Exception('No side for loops')
+                for loop in connector.loops:
+                    dot.edge(f'{connector.name}:p{loop[0]}{loop_side}:{loop_dir}',
+                             f'{connector.name}:p{loop[1]}{loop_side}:{loop_dir}')
 
         for _, cable in self.cables.items():
 
@@ -170,40 +163,50 @@ class Harness:
                           f'{cable.length} m' if cable.length > 0 else '']
             attributes = list(filter(None, attributes))
 
-            html = '<table border="0" cellspacing="0" cellpadding="0"><tr><td>'  # main table
+            html = '<table border="0" cellspacing="0" cellpadding="0">'  # main table
 
-            html = f'{html}<table border="0" cellspacing="0" cellpadding="3" cellborder="1">'  # name+attributes table
-            if cable.show_name:
-                html = f'{html}<tr><td colspan="{len(attributes)}">{cable.name}</td></tr>'
-            if(len(identification) > 0):  # print an identification row if values specified
-                html = f'{html}<tr><td colspan="{len(attributes)}" cellpadding="0"><table border="0" cellspacing="0" cellpadding="3" cellborder="1"><tr>'
-                for attrib in identification[0:-1]:
-                    html = f'{html}<td sides="R">{attrib}</td>' # all columns except last have a border on the right (sides="R")
-                if len(identification) > 0:
-                    html = f'{html}<td border="0">{identification[-1]}</td>' # last column has no border on the right because the enclosing table borders it
-                html = f'{html}</tr></table></td></tr>'  # end identification row
-            html = f'{html}<tr>'  # attribute row
-            for attrib in attributes:
-                html = f'{html}<td balign="left">{attrib}</td>'
-            html = f'{html}</tr>'  # attribute row
-            html = f'{html}</table></td></tr>'  # name+attributes table
+            if cable.show_name or len(attributes) > 0:
+                html = f'{html}<tr><td><table border="0" cellspacing="0" cellpadding="3" cellborder="1">'  # name+attributes table
+                if cable.show_name:
+                    html = f'{html}<tr><td colspan="{max(len(attributes), 1)}">{cable.name}</td></tr>'
+                if(len(identification) > 0):  # print an identification row if values specified
+                    html = f'{html}<tr><td colspan="{len(attributes)}" cellpadding="0"><table border="0" cellspacing="0" cellpadding="3" cellborder="1"><tr>'
+                    for attrib in identification[0:-1]:
+                        html = f'{html}<td sides="R">{attrib}</td>' # all columns except last have a border on the right (sides="R")
+                    if len(identification) > 0:
+                        html = f'{html}<td border="0">{identification[-1]}</td>' # last column has no border on the right because the enclosing table borders it
+                    html = f'{html}</tr></table></td></tr>'  # end identification row
+                if(len(attributes) > 0):
+                    html = f'{html}<tr>'  # attribute row
+                    for attrib in attributes:
+                        html = f'{html}<td balign="left">{attrib}</td>'
+                    html = f'{html}</tr>'  # attribute row
+                html = f'{html}</table></td></tr>'  # name+attributes table
 
             html = f'{html}<tr><td>&nbsp;</td></tr>'  # spacer between attributes and wires
 
             html = f'{html}<tr><td><table border="0" cellspacing="0" cellborder="0">'  # conductor table
 
-            for i, connection in enumerate(cable.colors, 1):
+            # determine if there are double- or triple-colored wires;
+            # if so, pad single-color wires to make all wires of equal thickness
+            colorlengths = list(map(len, cable.colors))
+            pad = 4 in colorlengths or 6 in colorlengths
+
+            for i, connection_color in enumerate(cable.colors, 1):
                 p = []
                 p.append(f'<!-- {i}_in -->')
-                p.append(wv_colors.translate_color(connection, self.color_mode))
+                p.append(wv_colors.translate_color(connection_color, self.color_mode))
                 p.append(f'<!-- {i}_out -->')
                 html = f'{html}<tr>'
                 for bla in p:
                     html = f'{html}<td>{bla}</td>'
                 html = f'{html}</tr>'
-                bgcolor = wv_colors.translate_color(connection, 'hex')
-                bgcolor = bgcolor if bgcolor != '' else '#ffffff'
-                html = f'{html}<tr><td colspan="{len(p)}" cellpadding="0" height="6" bgcolor="{bgcolor}" border="2" sides="tb" port="w{i}"></td></tr>'
+
+                bgcolors = ['#000000'] + get_color_hex(connection_color, pad=pad) + ['#000000']
+                html = f'{html}<tr><td colspan="{len(p)}" border="0" cellspacing="0" cellpadding="0" port="w{i}" height="{(2 * len(bgcolors))}"><table cellspacing="0" cellborder="0" border = "0">'
+                for j, bgcolor in enumerate(bgcolors[::-1]):  # Reverse to match the curved wires when more than 2 colors
+                    html = f'{html}<tr><td colspan="{len(p)}" cellpadding="0" height="2" bgcolor="{bgcolor if bgcolor != "" else wv_colors.default_color}" border="0"></td></tr>'
+                html = html + '</table></td></tr>'
                 if(cable.category == 'bundle'):  # for bundles individual wires can have part information
                     # create a list of wire parameters
                     wireidentification = []
@@ -241,32 +244,26 @@ class Harness:
             html = f'{html}</table>'  # main table
 
             # connections
-            for connection in cable.connections:
-                if isinstance(connection.via_port, int):  # check if it's an actual wire and not a shield
-                    search_color = cable.colors[connection.via_port - 1]
-                    if search_color in wv_colors.color_hex:
-                        dot.attr('edge', color=f'#000000:{wv_colors.color_hex[search_color]}:#000000')
-                    else:  # color name not found
-                        dot.attr('edge', color='#000000:#ffffff:#000000')
+            for connection_color in cable.connections:
+                if isinstance(connection_color.via_port, int):  # check if it's an actual wire and not a shield
+                    dot.attr('edge', color=':'.join(['#000000'] + wv_colors.get_color_hex(cable.colors[connection_color.via_port - 1], pad=pad) + ['#000000']))
                 else:  # it's a shield connection
-                    dot.attr('edge', color='#000000')
-
-                if connection.from_port is not None:  # connect to left
-                    from_ferrule = self.connectors[connection.from_name].category == 'ferrule'
-                    port = f':p{connection.from_port}r' if not from_ferrule else ''
-                    code_left_1 = f'{connection.from_name}{port}:e'
-                    code_left_2 = f'{cable.name}:w{connection.via_port}:w'
+                    # shield is shown as a thin tinned wire
+                    dot.attr('edge', color=':'.join(['#000000', wv_colors.get_color_hex('SN', pad=False)[0], '#000000']))
+                if connection_color.from_port is not None:  # connect to left
+                    from_port = f':p{connection_color.from_port}r' if self.connectors[connection_color.from_name].style != 'simple' else ''
+                    code_left_1 = f'{connection_color.from_name}{from_port}:e'
+                    code_left_2 = f'{cable.name}:w{connection_color.via_port}:w'
                     dot.edge(code_left_1, code_left_2)
-                    from_string = f'{connection.from_name}:{connection.from_port}' if not from_ferrule else ''
-                    html = html.replace(f'<!-- {connection.via_port}_in -->', from_string)
-                if connection.to_port is not None:  # connect to right
-                    to_ferrule = self.connectors[connection.to_name].category == 'ferrule'
-                    code_right_1 = f'{cable.name}:w{connection.via_port}:e'
-                    to_port = f':p{connection.to_port}l' if not to_ferrule else ''
-                    code_right_2 = f'{connection.to_name}{to_port}:w'
+                    from_string = f'{connection_color.from_name}:{connection_color.from_port}' if self.connectors[connection_color.from_name].show_name else ''
+                    html = html.replace(f'<!-- {connection_color.via_port}_in -->', from_string)
+                if connection_color.to_port is not None:  # connect to right
+                    code_right_1 = f'{cable.name}:w{connection_color.via_port}:e'
+                    to_port = f':p{connection_color.to_port}l' if self.connectors[connection_color.to_name].style != 'simple' else ''
+                    code_right_2 = f'{connection_color.to_name}{to_port}:w'
                     dot.edge(code_right_1, code_right_2)
-                    to_string = f'{connection.to_name}:{connection.to_port}' if not to_ferrule else ''
-                    html = html.replace(f'<!-- {connection.via_port}_out -->', to_string)
+                    to_string = f'{connection_color.to_name}:{connection_color.to_port}' if self.connectors[connection_color.to_name].show_name else ''
+                    html = html.replace(f'<!-- {connection_color.via_port}_out -->', to_string)
 
             dot.node(cable.name, label=f'<{html}>', shape='box',
                      style='filled,dashed' if cable.category == 'bundle' else '', margin='0', fillcolor='white')
@@ -300,14 +297,19 @@ class Harness:
         graph.save(filename=f'{filename}.gv')
         # bom output
         bom_list = self.bom_list()
-        with open(f'{filename}.bom.tsv', 'w') as file:
+        with open_file_write(f'{filename}.bom.tsv') as file:
             file.write(tuplelist2tsv(bom_list))
         # HTML output
-        with open(f'{filename}.html', 'w') as file:
-            file.write('<html><body style="font-family:Arial">')
+        with open_file_write(f'{filename}.html') as file:
+            file.write('<!DOCTYPE html>\n')
+            file.write('<html><head><meta charset="UTF-8"></head><body style="font-family:Arial">')
 
             file.write('<h1>Diagram</h1>')
-            with open(f'{filename}.svg') as svg:
+            with open_file_read(f'{filename}.svg') as svg:
+                file.write(re.sub(
+                    '^<[?]xml [^?>]*[?]>[^<]*<!DOCTYPE [^>]*>',
+                    '<!-- XML and DOCTYPE declarations from SVG file removed -->',
+                    svg.read(1024), 1))
                 for svgdata in svg:
                     file.write(svgdata)
 
@@ -343,10 +345,10 @@ class Harness:
             designators.sort()
             conn_type = f', {remove_line_breaks(shared.type)}' if shared.type else ''
             conn_subtype = f', {remove_line_breaks(shared.subtype)}' if shared.subtype else ''
-            conn_pincount = f', {shared.pincount} pins' if shared.category != 'ferrule' else ''
+            conn_pincount = f', {shared.pincount} pins' if shared.style != 'simple' else ''
             conn_color = f', {shared.color}' if shared.color else ''
             name = f'Connector{conn_type}{conn_subtype}{conn_pincount}{conn_color}'
-            item = {'item': name, 'qty': len(designators), 'unit': '', 'designators': designators if shared.category != 'ferrule' else '',
+            item = {'item': name, 'qty': len(designators), 'unit': '', 'designators': designators if shared.show_name else '',
                     'manufacturer': shared.manufacturer, 'manufacturer part number': shared.manufacturer_part_number, 'internal part number': shared.internal_part_number}
             bom_connectors.append(item)
             bom_connectors = sorted(bom_connectors, key=lambda k: k['item'])  # https://stackoverflow.com/a/73050