+++ src/wireviz/wv_helper.py
@@ -111,3 +111,10 @@ def graphviz_line_breaks(inp):
 
 def remove_line_breaks(inp):
     return inp.replace('\n', ' ').rstrip() if isinstance(inp, str) else inp
+
+def open_file_read(filename):
+    # TODO: Intelligently determine encoding
+    return open(filename, 'r', encoding='UTF-8')
+
+def open_file_write(filename):
+    return open(filename, 'w', encoding='UTF-8')